Server/requirement.txt
Django
south
django-cors-headers
djangorestframework
MySQL-python
django-filter
drf-nested-routers
drf-extensions
django-oauth2-provider
qiniu
